feat: relocate from_id and to id methods to the parent class
